Fix change password
