+++ pinterest/repository/repository_methods.go
@@ -3,6 +3,7 @@ package repository
 import (
 	"database/sql"
 	"github.com/go-park-mail-ru/2019_2_Solar/pkg/consts"
+	"github.com/go-park-mail-ru/2019_2_Solar/pkg/functions"
 	"github.com/go-park-mail-ru/2019_2_Solar/pkg/models"
 	"github.com/lib/pq"
 	_ "github.com/lib/pq"
@@ -600,7 +601,14 @@ func (RS *ReposStruct) SelectIDUsernameEmailUser(username, email string) (Users 
 }
 
 func (RS *ReposStruct) UpdateUser(user models.User) (int, error) {
-	result, err := RS.DataBase.Exec(consts.UPDATEUserByID, user.Username, user.Name, user.Surname, user.Password, user.Email, user.Age, user.Status, user.ID)
+	// Crutch
+	salt, err := functions.GenSessionKey(10)
+	if err != nil {
+		return 0, err
+	}
+	hashPassword := functions.HashPassword(user.Password, salt)
+
+	result, err := RS.DataBase.Exec(consts.UPDATEUserByID, user.Username, user.Name, user.Surname, hashPassword, user.Email, user.Age, user.Status, user.ID, salt)
 	if err != nil {
 		return 0, err
 	}
