+++ pkg/consts/sqlquery.go
@@ -5,7 +5,7 @@ const (
 	SELECTUserIDUsernameEmailByUsernameOrEmail = "SELECT u.id, u.username, u.email from sunrise.user as u where u.username = $1 OR u.email = $2"
 	SELECTAllUsers                             = "SELECT * from sunrise.user"
 	UPDATEUserByID                             = "UPDATE sunrise_db.sunrise.user SET username = $1, name = $2, 	surname = $3," +
-		" hashpassword = $4,email = $5, age = $6, status = $7 where id = $8"
+		" hashpassword = $4,email = $5, age = $6, status = $7, salt = $9 where id = $8"
 	UPDATEUserAvatarDirByID = "UPDATE sunrise_db.sunrise.user SET avatardir = $1 where id = $2"
 	INSERTUser              = "INSERT INTO sunrise.user (username, email, hashpassword, salt, created_time)	values ($1,$2,$3,$4,$5) RETURNING id"
 	INSERTSession           = "INSERT INTO sunrise.usersession (userid, cookiesvalue, cookiesexpiration)	values ($1,$2,$3) RETURNING id"
